# reduce allocs and improve the render `WriteString`